requirements.txt
paramiko
nmap-python
pwntools
